pyproject.toml
[project]
name = "otus-mlflow-local"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3==1.36.23",
    "invoke==2.2.0",
    "ipykernel==6.29.5",
    "loguru==0.7.3",
    "mlflow==2.20.2",
    "pandas>=2.2.3",
    "python-dotenv==1.0.1",
    "scikit-learn==1.6.1",
]
